refactor (#2)

+++ src/image_processor.cpp
@@ -15,7 +15,7 @@ std::vector<uint8_t> BiLinearInterPolation(std::vector<uint8_t> &inputImage, Tga
     for (auto ix = 0; ix < tgaImage.tgaHeader_.height - 1; ix += 2) {
         for (auto iy = 0; iy < tgaImage.tgaHeader_.width - 1; iy += 2) {
             auto jump = (tgaImage.tgaHeader_.width * ix * tgaImage.pixelSize_) + iy * tgaImage.pixelSize_;
-            for (size_t ij = 0; ij < tgaImage.pixelSize_; ij++) {
+            for (uint32_t ij = 0; ij < tgaImage.pixelSize_; ij++) {
                 interpolatedPixel =
                         (inputImage[jump + ij] + inputImage[tgaImage.Seek(jump + ij, 'R')] +
                          inputImage[tgaImage.Seek(jump + ij, 'D')] +
